backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
flask-limiter==3.9.0
msal==1.31.0
requests==2.32.3
gunicorn
